+++ src/index.ts
@@ -5,6 +5,7 @@ export * from './RokuDeploy';
 export * from './util';
 export * from './RokuDeployOptions';
 export * from './Errors';
+export * from './DeviceInfo';
 
 //create a new static instance of RokuDeploy, and export those functions for backwards compatibility
 export const rokuDeploy = new RokuDeploy();